Have `State.forward` take a required `key` argument as opposed to an optional one
Summary:
Inside `Fixpoint` we always pass an explicit `key` to `State.forward`, and that's the only callsite of `State.forward` in production.

I suspect that the reason why we made `key` optional is to accommodate tests, which is not a good justification for changing prod behavior.

Let's make the `key` argument mandatory instead of optional, so inside `State.forward` we don't need to worry about an absent key.

Reviewed By: dkgi

Differential Revision: D19871396

fbshipit-source-id: 408d33895c18b616c811c97a93261da42f45c97e