Late Combination
Summary:
Just as we were able to get significant time and space wins from late instantiation (D19638896) we are able to get corresponding wins from postponing our combination of the constituent tables of an attribute table.

Currently we process common base classes (like `object`) once for every child class.  This is because in transitive lookups, we combine all of the parent tables before placing it in the cache.  This is oblivious of the fact that that's a lot of overlapping work, and lots of overlapping stored information.

This diff fixes that by splitting what was once `definition_attributes` off into `single_uninstantiated_attribute_table` and deferring the "combination" of these subtables until they're actually needed.

This also allows us to get even fancier by only fetching the tables (and thus adding the dependency edges) as needed, which for single attribute lookups can be significant.  For example, when I access attribute `a` of child class `C`, I never need to look at the attributes of base class `B` as long as `C` actually has the attribute I'm looking for.

This is the origin of all of the `Sequence` and `lazy` stuff inside of `uninstantiated_attribute_tables`

Reviewed By: grievejia

Differential Revision: D19850235

fbshipit-source-id: dc4bb6dfa73c204f8a0c9d444e6f60f042f62a89