Treat unannotated return type as `Any` instead of `Top`
Summary:
We were setting the return type of unannotated functions to be `Top`, which is a problematic notion because it ought to be semantically equivalent to `object` but we don't treat it as such. Phasing out the usage of `Top` should be one of our long-term goals.

In addition, according to PEP484,

> For a checked function, the default annotation for arguments and for the return type is Any.

So we'd better stick to the standard here.

Reviewed By: dkgi

Differential Revision: D19896082

fbshipit-source-id: 7768914164f01f30dde47edbf289a772278cd025