Add functionality to retrieve superclasses in query API
Summary:
Adding a `get_superclasses` function to the query API so that the visitor can properly identify where the attribute in question is first defined.

e.g. The attribute `a` is getting modified below, but not in `Base` where it is defined. It is important to recognize this and make sure to avoid incorrectly marking `a` as `Final` in `Base`.
```
class Base:
  def __init__(self, a:int) -> None:
    self.a: int = a

class Child(Base):
  def __init__(self, a:int) -> None:
    super(a)
  def modify_attr(self) -> None:
    self.a += 5
```

Reviewed By: dkgi

Differential Revision: D19843279

fbshipit-source-id: 7bdf8ce4800461ef2f4f5ddf76d47a41e33a5715