resolution_fixpoint lives outside of the fixpoint
Summary:
Context: Our resolution fixpoint getting computed properly was relying on the fixpoint
executing in the right order. If you happened to have the wrong branch analyzed first, your fixpoint
wouldn't include the right resolutions. This is a fundamental problem: The computation of the resolution
fixpoint happens to be part of the fixpoint, but it shouldn't be thought of as part of the state - it's simply
intermediate resolutions that we store.

This diff changes the representation to actually represent the resolution_fixpoint how it is - something
that happens to be computed during the fixpoint, but is not part of widen, less_or_equal, etc. We do so
by representing it as a hash map, and having a `to_read_only` interface that can be use to read from the computation - it
doesn't make sense to read from the resolution_fixpoint while it is getting computed.

This change also fixes a tricky pysa bug - see the test.

Reviewed By: grievejia

Differential Revision: D19892734

fbshipit-source-id: f165ca2b6f22ba1882e29687a4976043f13635ea