Do not register getitems with unknowns as aliases
Summary:
Ensure we're properly resolving expressions of the format `Class[expression]` where the expression resolves to something unknown. Previously, we were registering these as type aliases and messing up analysis later on.

Type.create tries to parse any `__getitem__` into a parametric type, and values in the parameters that are not valid types get filtered away in a later step. Abort if the parameters contain unknown, too.

Reviewed By: sinancepel

Differential Revision: D19900463

fbshipit-source-id: 91fbf76608f3d7a28f1b8da30d0bb52e9d1049ef