Persist loaded saved state
Summary:
It turns out that D19508840 and D19525720 interact in a surprising fashion.  Unbeknownst to us, hack shared memory "loads" a dependency table from saved state by:
* connecting to the sqlite database of the dependency table *only in the worker*
* saving the *filename* of the dependency table into shared memory
* in every worker
    * only upon the first read of the dependency table, start sqlite, pointed to that filename stored by the worker.

D19525720 started deleting the saved state files "after" loading.  As we know now, loading isn't a once-and-done process, but was appearing so, as long as we only ever read from the dependency table in the master.  This was true, until D19508840.  This diff parallelized the dependency collection process, thus making the workers start to read from the dependency table.  Because the file had been deleted, this would crash.

The testing we did on D19508840 didn't surface this issue because I only did small changes off of the loaded state, which fell below the policy for parallelism.

The fix is to give up on trying to not store saved states, instead trying to just store one copy of the saved state.  I accomplish this here by storing it in the log_directory.

Reviewed By: MaggieMoss

Differential Revision: D19956216

fbshipit-source-id: 7a562824a1c95a890570e97459502644e5aef6a2