Using preprocessing to show how namedtuple attributes are initialized
Summary:
Just like for D19980118, we were previously marking namedtuple attributes as uninitialized.  Unlike dataclasses, namedtuples are handled in preprocessing.  To keep things consistent it made the most sense to me to add the handling for this one into preprocessing instead of environment building.

This also seems to be handling some stuff that was being special cased in Pysa, we should try deleting that now and see if it still works.

Reviewed By: dkgi

Differential Revision: D19979356

fbshipit-source-id: 270be139a9f99b945b0d8d541ba7b3036fe3cb75