Simpler strategy to distinguish real from stubbed Anys
Summary:
The existing strategy for this requires the allow_primitives_from_empty_stubs flag, which we'd like to kill.  Luckily we already have an alternative implementation in annotatedBases we just need to expose

NB: In theory base_is_from_placeholder_stub could take an Expression.t.  However, to emphasize this is really for just class bases, and other things should go through higher powered tools, I left it as close as a base type as we have

Reviewed By: grievejia

Differential Revision: D19986110

fbshipit-source-id: 4fe33df690601b9dbbfb83d0e04b4dde6e9a1214