use `resolve_ignoring_optional` for call targets
Summary:
Generally, we over-approximate taint in Pysa, and part of this extends to method calls on optional attributes. From the perspective of Pyre, it's fine to stop once you hit an `Optional[T]`. However, for Pysa, we still want to catch potential flows of taint if `T.method` is called.

Towards this end, make our target resolution ignore optionals and resolve the base types.

Reviewed By: dkgi

Differential Revision: D20014325

fbshipit-source-id: 951dec50cb658a2cf9a6090099c623a9e24b3a4d