track type breadcrumbs for tuples as well
Summary: This is not a perfect solution - longer term, we probably want to actually destruct the tuple, and only transform the right access path. However, that's more expensive to do, and this unblocks us for now. I'm tracking that at T62615143, but it's a lot higher effort with non-concrete payoff.

Reviewed By: dkgi

Differential Revision: D20006445

fbshipit-source-id: d6673d8c07e131621755ea4b2d3f276effb99b57