Hide unnecessarily exposed APIs from `TypeCheck.State`
Summary:
We were exposing `TypeCheckState.forward_expression` and `TypeCheckState.forward_statement` just for testing purpose. This had to be done because we did not expose a good set of APIs to query the result of `forward_XXX` in the past. Hence we were forced to resort to fiddle with these implementation details.

Now that with all the previous diffs on the stack, we finally have APIs like `Resolution.resolve_expression` and `Resolution.resolve_statement` that are almost as powerful as their `forward_XXX` counterpart. The `forward_XXX` family of functions can therefore be hidden now.

Tests that were previously rely on the `forward_XXX` functions need to be redirected to the new `Resolution` APIs. Turns out that ~80% of this diff is about such test refactor :P

Reviewed By: dkgi

Differential Revision: D20051326

fbshipit-source-id: 5b95a3ffdfb8b4d43e5764b32aa8eca178ee573a