introduce `metaruns` and `metaruns_runs_assoc` tables
Summary:
The design we've arrived on for metaruns and metaruns <-> runs is the following: We create a separate metarun table per db, have a many-to-many assoc table between metaruns and runs. The scheduler job for each metarun will populate the meta run id, and the child jobs will fill out the assoc.

We keep a string key for meta run ids, as we want these to be, for instance, sandcastle nonces or the string 'latest'. We might want to add a separate fbid/dbid field in the future, and we haven't designed ourselves away from that possibility - an addition, if necessary, would be straightforward.

Reviewed By: fahndrich

Differential Revision: D20109155

fbshipit-source-id: 26352bf4ada99ec8d898234117ce07dd008c1322