Replace `emit_raw_error` with `emit_error`
Summary:
We have so many different ways of emitting errors in the type checker, which makes error-related refactoring painful. Let's redirect error-emitting logic to a common API instead.

I made the decision of using `emit_error` as this common API (`emit_raw_error` would be an equally valid choice but calling it tend to be more verbose). This diff replaces all usage of `emit_raw_error` with `emit_error`, and hide `emit_raw_error` inside `emit_error` so that we can't add more callsites to `emit_raw_error` in the future.

Reviewed By: dkgi

Differential Revision: D20102225

fbshipit-source-id: ad737b298b26cd939716135eb35160456a81cf60