Set initial resolution fixpoint at the end of `State.initial`
Summary:
We were previously setting the resolution fixpoint in the middle of `check_parameter_annotations`, which works but is rather weird (it depends on the assumptions that other checks that come after `check_parameter_annotations` do not mutate annotation stores).

Let's instead writes to resolution fixpoint at the end, to better reflect what resolution fixpoint is intended to hold.

Reviewed By: dkgi

Differential Revision: D20111218

fbshipit-source-id: 9d57f28f0b51a3dd6ae937625d18ce1af50ba835