Accessing fallback __getattr__ as a instance method rather than as a classmethod
Summary:
Because we were accessing it as a class method, we were getting the self type there, forcing us to call it with an additional argument (the class name itself as a reference).

However, this doesn't actually work, since the class name is of type Type[T], not T, so  it would normally be rejected.  However, unmarked self is currently parsed as being Top, so was being accepted.

The easy fix to this is not to give it an argument of type T, but just to access it the right way and have the self stripped off for us.

Reviewed By: dkgi

Differential Revision: D20132498

fbshipit-source-id: bbc471802718249435e5a9e514c2dded2e973bbd