Moving annotatedClass.fallback_attribute to resolution
Summary:
This is the last thing in annotatedclass, and it also doesn't really belong.  In general this function is kind of sketchy and probably just ought to be inlined to its one usage, but that would remove the existing (fairly robust) testing, so I'll kick that can by putting it in Resolution.

This finally allows us to kill AnnotatedClass

Reviewed By: dkgi, grievejia

Differential Revision: D20137701

fbshipit-source-id: fd122c0d095ae6e6d817cab21cab0cda6c8dd45b