Consistently inferring self/cls argument
Summary:
PEP 484 defines behavior that the first argument for methods of classes is only optionally annotated, and if not, should be inferred to take the parent class.
We currently do this by adding the annotation when type checking the body of the function.
However, this means that externally, this parameter is left unannotated, leading to unsoundnesses.
This is because even if a function is a method, it can still be called with the first parameter intact if its called on the class object rather than an instance.

This diff fixes this by consistently adding that first argument in the parse, ensuring that we always type check it consistently.

Reviewed By: dkgi

Differential Revision: D20132799

fbshipit-source-id: 7d0c306056c1496a4fa1531c89fc1b3197d40f5f