Type[T] <: AMetaclass if MetaclassOf[T] <: AMetaclass
Summary:
What it means to have a metaclass is precisely that you are affecting the superclasses of your metatype.  This diff adds that behavior to the type order.

This came up from experimenting with the `BoundMethod` idea, as one of our tests was failing when you called `Type[T].method_defined_in_metaclass_of_t()`, since the self type was expecting a `Metaclass`, and before this diff a `Type[T]` was not a `Metaclass`

Reviewed By: grievejia

Differential Revision: D20153240

fbshipit-source-id: 80e9c7dda4ec028deb8e87efcf1b8096d53cce1f