Unannotated parameters always get `Any` type regardless of the existence of default value
Summary:
According to PEP484, unannotated parameters always get the `Any` type. But currently we don't do it when those parameters are initialized with some default values, in which case we'll treat the parameters as the inferred type of the corresponding default values.

The first issue with our current behavior is, obviously, that we are deviating from what the standard says, which is not a good idea unless there's a strong justification. The second issue is that the inferred type might not always be a good guess of what the parameter type should be. Consider this example:

```
def foo(x = None) -> Something:
  if x:
    return something()
  else:
    return something_else()
```

Usually, the intention of the developer, when writing down `x = None`, is to allow the caller to select the two branches with the `x` argument, using the `something_else()` branch as the default. It does not mean that the type of `x` should be `None` -- which would make `pyre` emit the spurious error that the `something()` branch is dead.

Reviewed By: sinancepel

Differential Revision: D20195658

fbshipit-source-id: 3c1aa26d5290238dd9b694484cf9c5758a682901