Add an API to extract the type parameters from a given subtype
Summary:
Here's the usage scenario: we have a type `T` and a name `G`, and we want to know if `T` is a subtype `G[X]` for some `X` -- if so, we want to know what `X` is. For example,
- We see the expression `await foo()`. We compute the type of `foo()`, and we want to know what the type of the await expression is (i.e. view the type as `Awaitable[T]` and extract `T` from it)
- We see the expression `x for x in bar()`. We compute the type of `bar()`, and we want to know what the type of `x` is (i.e. view the type as `Iterable[T]` and extract `T` from it).

Here's how we do it in the past: we join the type with `G[Bottom]`, see if the result is `G[X]`, and return `X`. The issue with this approach is that it does not play well with `Any` (see T56720048).

The proper way of doing this is to rely on constraint solving. Sadly we did not have a good API for that in the past -- which is where this diff comes into play. I named the new API `GlobalResolution.extract_type_parameters`. It is a little bit more general than what was described above in the sense that it support those `G`s that have multiple type parameters.

I've only put a very minimal effort into this API. It has some weird behaviors with respect to `Top` and `Bottom` (e.g. you can't extract `Top`/`Bottom` from `List[Top]`/`List[Bottom]`). But still, I think this is better than what we had before.

Reviewed By: sinancepel

Differential Revision: D20163442

fbshipit-source-id: bad13b8bfdbbafa62584fb2b47b9d00a40c20cc0