Simplifying the join of the results of signature selection
Summary:
There's no reason why we need to bother with joining all of the parameters of the callables we get from signature select, since we only ever use the return type.
insisting on joining these callables to callables forces us to do extra work, and introduce a failure state that's unecessary.

Reviewed By: grievejia

Differential Revision: D20266914

fbshipit-source-id: 850d0c177cb981604d73f4845e7bb15e94ed8f16