simulate writes for property setters
Summary:
This diff makes us more accurately simulate effects for property setters. In general, it's not
scalable to keep effects models for each function, but similar to constructors, property setters
have a common enough pattern that a simple approach gets us much better results.

The general approach taken is this: When you see `a.property = x`, simulate this as
`a = A.property(a, x)`. This approach, the same as handling `x = C(args)` as `x = C.__init__(args)`,
requires us to return the taint corresponding to self. Once we have this set up, the actual
implementation is simple - just assign the entire taint returned from the property setter to `a`. Our
approach allows us to avoid having to do things like special case joins of taint for property setters, and lets
the infrastructure take care of everything.

Reviewed By: fahndrich

Differential Revision: D20285456

fbshipit-source-id: a7ff3ab248f7f641f54c30c40ad9b48767147aaf