Use state equality instead of leq for fixpoint convergence check
Summary:
When checking whether the fixpoint algorithm has converged or not, we rely on `State.less_or_equal`, which in turn tries to perform element-wise `RefinementUnit.less_or_equal`. Problem is, the scheme does not play well with `Any`.

Consider this example:

```
def bar() -> Optional[Any]: ...

def foo(flag: bool) -> None:
    if flag:
        x = bar()
    else:
        x = None
    reveal_type(x)
```

One would expect that we reveal `Optional[Any]`. But instead we actually reveal `None`.

The problem here is that, when we join `Optional[Any]` with `None` for `x`, we'll get `Optional[Any]`. But `Optional[Any]` is `less_or_equal` to  `None` (i.e. `Optional[Bottom]`), so Pyre thinks that the new state

```
x -> Optional[Any]
```

does not carry more info than the old state

```
x -> None
```

And the fixpoint algorithm would terminate prematurely. But this is not what we want -- we do prefer `Optional[Any]` to `Optional[Bottom]` here because the former tells us that Pyre is not sure about what the underlying type is, but the latter (erroneously) does not tell us so. Having `x` converged to `None` might not be a problem at first glance, but when you take other checks into account, it becomes more problematic:

```
def bar() -> Optional[Any]: ...

def foo(flag: bool) -> None:
    if flag:
        x = bar()
    else:
        x = None
    assert x  # Pyre wrongly complains that `x` has type `None` therefore assertion will always fail.
```

Also note that this is not just an issue with `Optional[Any]` -- any state that contains `Any` may be potentially affected by this.

The fix I propose here is to change the convergence criteria to rely on `RefinementUnit.equal` as opposed to `RefinementUnit.less_or_equal` -- this will correctly propagate those `Any`s during fixpoint and give us more reasonable resolutions in the end.

Reviewed By: sinancepel

Differential Revision: D20241971

fbshipit-source-id: 42f2fc2828cc0c586ff8cbd10373332f739db495