remove extra copies of finally from cfg
Summary:
The high level idea is the following insight: Exception handling in Python is glorified goto's, and what happens in a finally if you have a return/uncaught exception is the following: You push the return/raise to the stack, "goto" the finally's body, if there is a return/raise in the finally, it gets executed, but the flow returns to the return/raise after the body of the finally is executed. What this means from a type checking perspective is: We don't need to simulate the body of the finally in order to get accurate type checking info for the finally body - the return/raise statement would be executed with the previous context, so we can keep only 1 copy of the finally.

The only exception to this is the following case: Both the body of the try and all exception handlers return/raise, which means that we wouldn't properly simulate the body of the finally. This diff currently doesn't handle that case, but it's straightforward to code in support if we think that case will occur in practice.

Here's what the CFG structure looks like now:

* *The normal jumps are unaffected*, except for when you have a raise in the try block.
* The body's normal control flow goes to orelse. It's exceptional control flow goes to the dispatch.
* The dispatch flows to each handler + finally (for when no handler matches an exception).
* Each handler flows to finally. (Even if an exception gets re-raised, the error exit will be to the finally).
* The orelse flows to the finally.
* The finally block flows to the normal + error exits, IF there is no return. (This is consistent with the runtime).

Reviewed By: dkgi

Differential Revision: D20239067

fbshipit-source-id: 41096d1a85be7ac5a01fb5324d5160cd2a46e878