Don't hardcode python 3.6 for venv test
The quoted left portion allows spaces in the virtualenv path, but the
right unquoted allows the glob to find whatever python version is in use
(there should only be one).

I benchmarked this (7.6 seconds) against `--preserve-pythonpath` (22.4
seconds) so feel this is worth a little comprimise on readability of the
script.