Add `Final` annotations to immutable fields [3/n]
Summary:
Codemod to apply `Final` annotations to immutable fields in `fbcode` python projects.

This is an effort to drive adoption of Pyre's final attribute refinement feature.

For more context:

It is currently not typesafe to do the following
```
if x.optional_attr:
  expects_non_optional(x.optional_attr)
```
due to possible interleaving calls changing the value of `x.optional_attr`.

However if we can infer `x.optional_attr` to be `Final`, we know that interleaving calls will never be able to reassign this attribute, allowing us to properly refine on it. Therefore, enforcing the use of `Final` is paramount for final attribute refinement.

Reviewed By: mrkmndz

Differential Revision: D20319511

fbshipit-source-id: 37b8d767d870b3a5e6ef89503e5fd4c3b650afb8