Don't track callgraph edges to protocol methods
Summary:
# TL:DR
Callgraph edges to protocol methods aren't being used right now, and make moving from `implicit` to `BoundMethod` difficult.  This diff removes them.

# The long version

The callgraph is used for taint analysis as well as some external consumers.  In general, the callgraph is useful when it connects functions with actual implementations, so we can correctly schedule taint analysis etc.
We are currently adding edges for all calls, including those which are of protocol methods, which have no implementations.

This is not a problem as is, but poses a problem for moving towards the `BoundMethod` type.  We'd like to no longer need the implicit field and the practice of altering the callable to remove its first argument and replace it with a `BoundMethod` type which encodes the actual runtime behavior of bundling the underlying callable with the "implicit argument", and only using it when the bound method is actually called.

One of the parts of the runtime interface of the `BoundMethod` type is the `__func__` property, which exposes the underlying callable.  This means that in order for one bound method to substitute for another, the underlying function must also be substitutable.  However this will not be the case for protocols.
```
class P(Protocol):
   # implicitly this is Callable[[P, int], str]
   def method(self, x: int) -> str: ...

class I:
   def other_method(self) -> str:
      return "A"
   # implicitly this is Callable[[I, int], str]
   def method(self, x: int) -> str:
      return self.other_method()
```
We'd like for I to implement P, but `Callable[[I, int], str]` is not less than `Callable[[P, int], str]`.

Concretely that means that `instance_of_p.method.__func__(other_instance_of_p, 7)` may fail, since if `instance_of_p` is actually an `I` at runtime, and `other_instance_of_p` is something else that doesn't have a `other_method`, that attribute access will fail.

This also means that accessing attributes of meta-types of protocols is also unsound (I'll address that in another diff).

The sound way to avoid this is to leave the behavior of attributes of protocols "as-is", as in having `instance_of_p.method` be a `Callable[[int], str]` rather than a `BoundMethod[Callable[[P, int], str], P]`.  This is sound because all of the potential implementors `BoundMethod`s are in fact those kinds of callables.  However, once we've removed implicit, that plain callable can't be used to make a callgraph edge.

To avoid doing a behavior change in a later refactor, I've bit the bullet up front in this diff, just directly making those callables what they will eventually have to be.

# The alternative

The alternative I see to this (other than just leaving in `implicit` and not doing the `BoundMethod` approach at all) is creating another kind of `BoundMethod` that doesn't expose the underlying callable at all, like `BoundProtocolMethod`.  In order to make this work we'd need to put in some special casing into the backend to allow for `BoundMethod[Callable[[I, int], str], I] <: ProtocolBoundMethod[Callable[[P, int], str], P]`.  This kind of special casing is completely avoidable with the approach here, since we can just treat `BoundMethod` like any other `__call__`able class and just special case the attribute lookup, which is a great deal cleaner.

Reviewed By: dkgi

Differential Revision: D20369134

fbshipit-source-id: 98580c9a60bfb83273e07936b54e182bbc2cdaa3