Assume comment lines following a fixme are part of fixme
Summary:
Responding to report that pyre upgrade only removed part of a fixme. Turns out this is because the pyre upgrade script will always add line continuations with a double space, like:

```
# pyre-fixme[5]: blah blah
#  blah blah
```

When deleting, it also assumes that pyre fixme comment continuations follow this pattern.

I believe this was put in place so we didn't accidentally delete unrelated comments that are stacked, but let me change it up to (1) make sure we're adding fixmes below any pre-existing comments, and (2) assume we are free to delete lines of comment until the next live line of code when removing an ignore.

Seems like a better approach because multi-line comments are often added manually by users, who won't do any fancy indenting.

Reviewed By: pradeep90

Differential Revision: D20397956

fbshipit-source-id: afc00644fc41d14ea01ff0ec17f5bc7082ddc0e9