Stop logging in _send_and_handle_socket_request once response is read
Summary:
The _read_stderr call in _call_client is enabled and disabled via _call_client_terminated.  However, the call in _send_and_handle_socket_request is not, leading to log messages sometimes confusingly appearing after the list of errors.

By doing the same approach in this use case, we avoid this behavior.

NOTE: In both cases there is a TOCTOU bug in that this flag is read, then the log message is printed.  However, due to the GIL I think its pretty unlikely that this happens, and given that this is purely cosmetic, it doesn't seem worth the locking overhead to fix.

Reviewed By: dkgi

Differential Revision: D20435699

fbshipit-source-id: f931a2aed6319acdea2957f8faf9ac3f238bfc4e