Splitting Callable
Summary:
Callables need to be splittable if we're going to treat them like a normal class for the purpose of attribute access (as we do in the next diff).  However doing so reveals that we're currently suppressing attribute assignment errors with this lack of splitting.

This diff makes that special case explicit and tags it with a task to follow up on.

Reviewed By: sinancepel

Differential Revision: D20445605

fbshipit-source-id: e35c6193d6fc71b8e69befd8241bb3cbe2fe6f47