Remove error callable special cases
Summary:
This is the follow up promised from D20393912: Treating Callable consistently with other classes when it comes to error suppression and filtering.

Concretely this means filtering out attribute access errors from callables that contain Top or a Mock in their signature.  Both of these behaviors aren't optimal in general, but it doesn't make sense to apply those principles unevenly.

Reviewed By: sinancepel

Differential Revision: D20448896

fbshipit-source-id: a3c183f9a41de7e9bd06ff0e7073679fc5342bbf