Add warnings and retry if stdlib not found in typeshed tar
Summary:
We should probably be dumping the subdirectories inside the given typeshed location when we find that it doesn't contain typeshed as we expect in the client.

Fiddled around with logging what we are calling `shutil.move` on, and cannot identify any strange behavior that may point to the file move (or move conflicts) that could have caused stdlib to go missing. We are moving temp/stdlib into .pyre/resource_cache/#/typeshed, and if stdlib already exists from a concurrent run of pyre, the move will simply throw a file already exists exception (which is caught and logged) and continue on as expected. We already verify that the resource cache typeshed directory exists or is created.

Starting to suspect that it may be missing when we untar the file we've downloaded from dewey.

To test this, have added a check to take a look in the freshly untarred typeshed and see if it contains stdlib as we expect. If not, print more information about what it does contain, and retry the download.

Let's roll out a client version soon-ish, and see if this fixes/reduces the issue, or at least has enough logging to pinpoint whether stdlib is not being downloaded at all, or somehow being lost along the way.

Reviewed By: pradeep90

Differential Revision: D20452783

fbshipit-source-id: 0324197d72ecefbb3c502c75afc41290c888a000