Factor out stream logging logic to avoid reintroducing trailing logs
Summary:
The bug fixed by D20435699 was a result of the enabling-disabling of logging logic being separate from the logging logic itself.

To avoid this in the future, I factored out that into a contextmanager, which ensures that we never forget to disable it

NOTE: There is one tiny behavior change, the handshake now takes place while the tail is running, rather than before.  This means that if that makes the server emit a log, it will now show up in the client log/console output.  In practice I don't think it actually does, but worth mentioning.

Reviewed By: dkgi

Differential Revision: D20436206

fbshipit-source-id: 5f013b285eeeabce0cea784e012ce6cee2f0f3ea