tailing logs for queries as well
Summary:
As noted on D20436206 theres no reason why we can't tail the server.stdout for the query logs as well.  In fact it makes things a bit simpler to do so unconditionally.

NOTE: This affords us the opportunity to undo the behavior change from D20436206, and move the handshake outside of the tailed region.  However, I think that covering that is a net positive, and I see no reason to take it out.

Reviewed By: sinancepel

Differential Revision: D20436819

fbshipit-source-id: bb159358f5cc10a436795679b4f78d9385c44d24