Exclude taint annotations from equality of models
Summary: This diff fixes an inconsistent state where we were including some, but not all, taint annotations in the meaning of equality for a model. This diff removes all taint from the equality operation, so that models can be compared based on what they represent, not the annotations they happen to conation. This will make using `FilteredSourceGenerator` more reliable.

Reviewed By: sinancepel

Differential Revision: D20408377

fbshipit-source-id: 1d3500fa1dfbc7ec249749ddb79bb8d35d5575c9