Avoid joining parameters to Bottoms
Summary:
Joining Contravariant[int] and Contravariant[str] to Contravariant[$bottom] ends up losing us a lot of information.  Furthermore, it causes problems in that bottom is special cased to not be compared in parametric less or equal checks, which means that explicit annotations are not checked against these types.

Big picture, we should try to have the same policy for Any, Top and Bottom: the only time we can create one is in the course of creating an error.

Reviewed By: sinancepel

Differential Revision: D20473400

fbshipit-source-id: eec279fc3bda85bc42a9116356f43b0d70d93a77