Introducing `BoundMethod`
Summary:
# Introduction
This diff is the very beginning of actually implementing a significant restructure of our method handling code: BoundMethod.

The core idea is to encode the `bound method` object which exists in the runtime into the type system.  Currently on attribute access we special case "methods" (things explicitly written as a `def`) as having their types modified on access through an instance.

# The status quo
Doing this in this artificial necessitates kludges like the `implicit` record member in callables to recover their full signature when we need it.  However, we don't implement this consistently enough to even include that member in our `equal` function for `Type.Callable.t`s.

It also requires doing a sort of hacky partial signature selection in order to handle the `TSelf` pattern.  This works for some limited cases but doesn't work for more complex kinds of parameters or with overloads or with more than one type variable in the argument.

Most critically, the existing implementation massively complicates any efforts to improve our handling of descriptors or decorators.  We'd like to implement decorators as an attribute lookup of `__get__` on types, but that only works if what the descriptor does can be encoded into a type.  `BoundMethod` is exactly that type.

Once we have this kind of descriptor working, we can easily handle decorators that return anything more complicated than a callable, one of the most frequent requests I've gotten from users.  With this, defines will just be class-level assignments of a particular type, allowing for all of the behavior changing decorators (`classmethod`, `staticmethod`, `property`) to be supported directly.

# This diff
Concretely this particular diff starts the ball rolling by including the `BoundMethod` class itself as a class in the built-in namespace.  This class has two generic parameters, the first for the callable, and the second for the self argument.  The only behavior that this class implements is a `__call__` attribute, which is defined by hooking into the existing `partial_apply_self` logic.

Doing this at the attribute level instantly gets us support as any other callable class, allowing us to work in the type order, and in signature selection, etc.

# Rejected Alternatives

One approach I initially thought would work was encoding the descriptor for functions as
```
    overload
    def __get__(
        self: Callable[TParams, TReturn], obj: None, objtype: object = None
    ) -> Callable[TParams, TReturn]:
        ...

    overload
    def __get__(
        self: Callable[Concatenate[TSelf, TParams], TReturn],
        obj: TSelf,
        objtype: object = None,
    ) -> Callable[TParams, TReturn]:
        ...
```

However, I came to realize that TParams cannot and should not be propagating local generics.  This also seems to be the standard for things like this, as TypeScript's ParametersOf also refuses to work on generic functions.

This means to support things like
```
class Foo:
  def generic_method(self, x: T) -> T: ...
```
we need something that's specially tailored to the behavior of the runtime, and can rely on it not being evil.

Reviewed By: grievejia

Differential Revision: D20267583

fbshipit-source-id: cce348a7f0351d52f44a8f33311aab7b82140f57