Reordering method special cases to be before partial application of self
Summary:
There is one more minor cleanup we need to do before embarking on a series of direct preparations for enabling BoundMethod.  We have a whole slew of special cases in instantiate_attribute for methods, which is fine: not everything (for example TypedDicts) can be encoded into the type system.
However, we'd like for all of that to precede our partial application of self, and eventually, hitting our descriptor logic.

This has the benefit of making these methods act much more consistently (i.e. TypedDictClass.__init__ now correctly has a self argument).

Reviewed By: grievejia

Differential Revision: D20484646

fbshipit-source-id: 190e1359019d8ab965446cded2d69df1e78516e9