Accepting __call__s that are BoundMethods
Summary:
This diff is part of a series designed to make the diff enabling BoundMethods as painless to review as possible.

This one allows for classes which have a `__call__` attribute thats a BoundMethod, not a Callable, to subclass a Callable.
After activation, nearly all callable classes will be this way.

Reviewed By: sinancepel

Differential Revision: D20485107

fbshipit-source-id: 0fc690c89a77fd6a560a86141927c80bc3673559