Replace `State.t` with `Resolution.t` in `State.initial`
Summary:
With D20370397 in, a state now just contains a resolution, a bottom flag, and  two global maps `resolution_fixpoint` and `error_map`.

With respect to expression-local type checking, only the resolution is relevant. With respect to statement-local type checking, only the resolution and the bottom flag are relevant. Therefore, we can further shrink the type checking APIs and have them operate on `Resolution.t` instead of `State.t`, eliminating the need of having to prepare for irrelevant info (e.g. two global maps) when invoking those APIs.

Turns out these refactoring are surprisingly easy with the help of OCaml type system. This diff changes `State.init` such that it passes `Resolution.t` around instead of `State.t` around. External interface of `State.init` stays the same -- so the change is completely local and straightforward.

Reviewed By: dkgi

Differential Revision: D20502443

fbshipit-source-id: 819eae65f0fdbd49e4d8ad2750372a1822431b3d